api/requirements.txt
firecrawl-py
openai
fastapi
uvicorn[standard]
slowapi